+++ CHANGES.rst
@@ -1,6 +1,10 @@
 Changes
 =======
 
+Unreleased
+-----------
+- Add `custom_model_name` parameter to the constructor of `HistoricalRecords` (gh-451)
+
 2.5.1 (2018-10-19)
 ------------------
 - Add `'+'` as the `history_type` for each instance in `bulk_history_create` (gh-449)