+++ CHANGES.rst
@@ -2,8 +2,9 @@ Changes
 =======
 
 Unreleased
------------
+----------
 - Add `custom_model_name` parameter to the constructor of `HistoricalRecords` (gh-451)
+- Fix header on history pages when custom site_header is used (gh-448)
 
 2.5.1 (2018-10-19)
 ------------------